fix(website): SDLC-Split-Nachlauf — G-CQ02-Duplikat, Goals und design-system ins richtige Target [T002639]
Nachlauf zu T002624 (ADR-006 Etappe 1). Alle drei Befunde stammen aus einer
transitiven Erreichbarkeitsanalyse (pages/sdlc/** gegen die uebrigen pages/**),
nicht aus Stichproben.

1. G-CQ02-Gate lag doppelt vor. code-quality.bats trug das archivierte Batch-1-
   Gate (Schwelle 373), g-cq02-any-types.bats die Nachfolgestufe (200).
   'monitoring.ts <= 2' war wortgleich doppelt — ein Fix an einer Stelle liess
   die andere rot und kostete beim Split eine CI-Runde. 'catch-blocks' war hier
   die schwaechere Variante (ohne 'catch (error: any)'), und das 373er-Gate kann
   seit der 200er-Stufe nicht mehr eigenstaendig anschlagen: es ist in der
   strengeren Schwelle logisch enthalten. Konsolidiert auf die strengere Datei;
   der Kopf von code-quality.bats begruendet die Abloesung der archivierten REQ-7.

2. Die Goals-Dateien wurden ausschliesslich von pages/sdlc/repohealth.astro
   erreicht, lagen aber im gemeinsamen Bereich. GoalsDashboard.svelte nach
   components/sdlc/, goals-data.ts und die generierte JSON nach lib/sdlc/;
   gen-goals-data.mjs schreibt an den neuen Ort.

3. pages/factory/design-system.astro lag ausserhalb von pages/sdlc/ und zog
   damit vier SDLC-Komponenten plus types.ts in den PROD-Build. Der Ein-Datei-
   Umzug beseitigt vier der sieben gemessenen Leckagen. pages/factory/ und
   pages/sdlc/ haben dieselbe Tiefe, daher blieben alle Importe gueltig.

Verifikation: BUILD_TARGET=prod und BUILD_TARGET=sdlc bauen beide fehlerfrei,
goals-data-Vitest 15/15, betroffene BATS 53/53.
