+++ docs/code-quality/repo-index.json
@@ -1118,7 +1118,6 @@
         "website/src/components/DevStatusTabs.svelte",
         "website/src/components/FAQ.svelte",
         "website/src/components/Footer.astro",
-        "website/src/components/GoalsDashboard.svelte",
         "website/src/components/Hero.svelte",
         "website/src/components/LanguageSwitcher.svelte",
         "website/src/components/MediaviewerPanel.svelte",
@@ -1363,6 +1362,7 @@
         "website/src/components/portal/WorkflowStatusMinimap.svelte",
         "website/src/components/sdlc/FactoryFloor.svelte",
         "website/src/components/sdlc/FactoryFloorLane.svelte",
+        "website/src/components/sdlc/GoalsDashboard.svelte",
         "website/src/components/sdlc/factory/AnalyticsWindowFilter.svelte",
         "website/src/components/sdlc/factory/AttentionStrip.svelte",
         "website/src/components/sdlc/factory/AwaitingDeployLane.svelte",
@@ -1682,9 +1682,6 @@
         "website/src/lib/generate-3d-pipeline.test.ts",
         "website/src/lib/generation-jobs.test.ts",
         "website/src/lib/generation-jobs.ts",
-        "website/src/lib/goals-data.generated.json",
-        "website/src/lib/goals-data.test.ts",
-        "website/src/lib/goals-data.ts",
         "website/src/lib/graph-utils.test.ts",
         "website/src/lib/graph-utils.ts",
         "website/src/lib/help-videos.test.ts",
@@ -1864,6 +1861,9 @@
         "website/src/lib/sdlc/factory-observability.ts",
         "website/src/lib/sdlc/github-ci.test.ts",
         "website/src/lib/sdlc/github-ci.ts",
+        "website/src/lib/sdlc/goals-data.generated.json",
+        "website/src/lib/sdlc/goals-data.test.ts",
+        "website/src/lib/sdlc/goals-data.ts",
         "website/src/lib/sdlc/k8s.test.ts",
         "website/src/lib/sdlc/k8s.ts",
         "website/src/lib/sdlc/ki-config-db.test.ts",
@@ -2421,7 +2421,6 @@
         "website/src/pages/en/stripe/success.astro",
         "website/src/pages/en/termin.astro",
         "website/src/pages/en/ueber-mich.astro",
-        "website/src/pages/factory/design-system.astro",
         "website/src/pages/impressum.astro",
         "website/src/pages/index.astro",
         "website/src/pages/kontakt.astro",
@@ -2583,6 +2582,7 @@
         "website/src/pages/sdlc/architektur.astro",
         "website/src/pages/sdlc/bugs.astro",
         "website/src/pages/sdlc/cockpit.astro",
+        "website/src/pages/sdlc/design-system.astro",
         "website/src/pages/sdlc/ki-konfiguration.astro",
         "website/src/pages/sdlc/observability.astro",
         "website/src/pages/sdlc/pipeline.astro",
