+++ website/src/lib/sdlc/goals-data.generated.json
