+++ website/src/lib/sdlc/goals-data.ts
