+++ website/src/pages/sdlc/design-system.astro
