+++ website/src/pages/sdlc/repohealth.astro
@@ -3,7 +3,7 @@ export const prerender = false;
 
 import AdminLayout from '../../layouts/AdminLayout.astro';
 import { getSession, getLoginUrl, isAdmin } from '../../lib/auth';
-import GoalsDashboard from '../../components/GoalsDashboard.svelte';
+import GoalsDashboard from '../../components/sdlc/GoalsDashboard.svelte';
 
 const session = await getSession(Astro.request.headers.get('cookie'));
 if (!session) return Astro.redirect(getLoginUrl(Astro.url.pathname));
